# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'BABEL v1'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /v1/smartolt/onu/status:
    get:
      summary: 'Get ONU Status'
      operationId: getONUStatus
      description: 'This endpoint allows you to get the status of an ONU.'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer BEARER_TOKEN'
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: success
                  message: 'ONU Status retrieved successfully.'
                  detail:
                    status: true
                    onu_status: Online
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'ONU Status retrieved successfully.'
                  detail:
                    type: object
                    properties:
                      status:
                        type: boolean
                        example: true
                      onu_status:
                        type: string
                        example: Online
        400:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Invalid data.'
                  detail:
                    onu_sn:
                      - 'The onu sn field is required.'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Invalid data.'
                  detail:
                    type: object
                    properties:
                      onu_sn:
                        type: array
                        example:
                          - 'The onu sn field is required.'
                        items:
                          type: string
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'ONU Status not found.'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'ONU Status not found.'
        500:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'An error occurred while retrieving ONU Status.'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'An error occurred while retrieving ONU Status.'
      tags:
        - 'SmartOLT Controller'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                onu_sn:
                  type: string
                  description: 'The ONU serial number.'
                  example: HWTCA7A1236A
              required:
                - onu_sn
  /v1/smartolt/onu/reboot:
    post:
      summary: 'Reboot ONU'
      operationId: rebootONU
      description: 'This endpoint allows you to reboot an ONU by its serial number.'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer BEARER_TOKEN'
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: success
                  message: 'ONU rebooted successfully.'
                  detail:
                    status: true
                    response: 'Device reboot command sent'
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'ONU rebooted successfully.'
                  detail:
                    type: object
                    properties:
                      status:
                        type: boolean
                        example: true
                      response:
                        type: string
                        example: 'Device reboot command sent'
        400:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Invalid data.'
                  detail:
                    onu_sn:
                      - 'The onu sn field is required.'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Invalid data.'
                  detail:
                    type: object
                    properties:
                      onu_sn:
                        type: array
                        example:
                          - 'The onu sn field is required.'
                        items:
                          type: string
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'ONU Status not found.'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'ONU Status not found.'
        500:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'An error occurred while retrieving ONU Status.'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'An error occurred while retrieving ONU Status.'
      tags:
        - 'SmartOLT Controller'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                onu_sn:
                  type: string
                  description: 'The ONU serial number.'
                  example: HWTCA7A1236A
              required:
                - onu_sn
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []